Move job into annotation module
